mv GenerateNominations to library
