+++ src/library/SetupSteps/ChooseNominations.tsx
@@ -4,28 +4,28 @@
 import { useConnect } from 'contexts/Connect';
 import { useUi } from 'contexts/UI';
 import { useApi } from 'contexts/Api';
-import { SetupType } from 'contexts/UI/types';
 import { Header } from 'library/SetupSteps/Header';
 import { Footer } from 'library/SetupSteps/Footer';
 import { MotionContainer } from 'library/SetupSteps/MotionContainer';
-import { GenerateNominations } from '../GenerateNominations';
-import { ChooseNominationsProps } from '../types';
+import { GenerateNominations } from './GenerateNominations';
+import { ChooseNominationsProps } from './types';
 
 export const ChooseNominations = (props: ChooseNominationsProps) => {
+  const { batchKey, setupType, section } = props;
+
   const { consts } = useApi();
-  const { section } = props;
   const { activeAccount } = useConnect();
   const { getSetupProgress, setActiveAccountSetup } = useUi();
-  const setup = getSetupProgress(SetupType.Stake, activeAccount);
+  const setup = getSetupProgress(setupType, activeAccount);
   const { maxNominations } = consts;
 
   const setterFn = () => {
-    return getSetupProgress(SetupType.Stake, activeAccount);
+    return getSetupProgress(setupType, activeAccount);
   };
 
   // handler for updating bond
   const handleSetupUpdate = (value: any) => {
-    setActiveAccountSetup(SetupType.Stake, value);
+    setActiveAccountSetup(setupType, value);
   };
 
   return (
@@ -36,7 +36,7 @@ export const ChooseNominations = (props: ChooseNominationsProps) => {
         title="Nominate"
         assistantPage="stake"
         assistantKey="Nominating"
-        setupType={SetupType.Stake}
+        setupType={setupType}
       />
       <MotionContainer thisSection={section} activeSection={setup.section}>
         <div style={{ marginTop: '0.5rem' }}>
@@ -45,7 +45,7 @@ export const ChooseNominations = (props: ChooseNominationsProps) => {
             nominations automatically or add a selection from your favourites.
           </h4>
           <GenerateNominations
-            batchKey="generate_nominations_inactive"
+            batchKey={batchKey}
             setters={[
               {
                 current: {
@@ -58,10 +58,7 @@ export const ChooseNominations = (props: ChooseNominationsProps) => {
             nominations={setup.nominations}
           />
         </div>
-        <Footer
-          complete={setup.nominations.length > 0}
-          setupType={SetupType.Stake}
-        />
+        <Footer complete={setup.nominations.length > 0} setupType={setupType} />
       </MotionContainer>
     </>
   );
