+++ src/library/SetupSteps/types.ts
@@ -3,6 +3,23 @@
 
 import { SetupType } from 'contexts/UI/types';
 
+export interface ChooseNominationsProps {
+  batchKey: string;
+  setupType: SetupType;
+  section: number;
+}
+
+export interface FooterProps {
+  complete: boolean;
+  setupType: SetupType;
+}
+
+export interface GenerateNominationsInnerProps {
+  setters: Array<any>;
+  nominations: string[];
+  batchKey: string;
+}
+
 export interface HeaderProps {
   title?: string;
   assistantPage?: string;
@@ -12,7 +29,4 @@ export interface HeaderProps {
   setupType: SetupType;
 }
 
-export interface FooterProps {
-  complete: boolean;
-  setupType: SetupType;
-}
+export type Nominations = string[];
