+++ src/pages/Stake/types.ts
@@ -1,19 +1,7 @@
 // Copyright 2022 @paritytech/polkadot-staking-dashboard authors & contributors
 // SPDX-License-Identifier: Apache-2.0
 
-export interface GenerateNominationsInnerProps {
-  setters: Array<any>;
-  nominations: string[];
-  batchKey: string;
-}
-
-export type Nominations = string[];
-
 export interface SetControllerProps {
-  section: number;
-}
-
-export interface ChooseNominationsProps {
   section: number;
 }
